♻️ Refactor: Change folder structure
