--- server/src/app.js
@@ -1,39 +1,0 @@
-const express = require("express");
-const cookieParser = require("cookie-parser");
-const cors = require("cors");
-const morgan = require("morgan");
-
-const globalErrorHandler = require("./middlewares/errorMiddleware");
-const authRouter = require("./routes/authRoutes");
-const userRouter = require("./routes/userRoutes");
-const artistRouter = require("./routes/artistRoutes");
-const songRouter = require("./routes/songRoutes");
-const playlistRouter = require("./routes/playlistRoutes");
-const searchRouter = require("./routes/searchRoutes");
-
-const app = express();
-
-// Middlewares
-app.use(
-  cors({
-    origin: process.env.CLIENT_URL,
-    credentials: true,
-  }),
-);
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-app.use(cookieParser());
-app.use(morgan("tiny"));
-
-// Routes
-app.use("/api/auth", authRouter);
-app.use("/api/users", userRouter);
-app.use("/api/artists", artistRouter);
-app.use("/api/songs", songRouter);
-app.use("/api/playlists", playlistRouter);
-app.use("/api/search", searchRouter);
-
-// Error handling middleware
-app.use(globalErrorHandler);
-
-module.exports = app;
