+++ server/test/helpers/createUsersAndLogin.js
@@ -1,5 +1,5 @@
 const request = require("supertest");
-const app = require("../../src/app");
+const app = require("../../src/config/app.config");
 const User = require("../../src/models/userModel");
 const {testUsers} = require("../testData");
