+++ server/test/helpers/signupAndLoginUser.js
@@ -1,5 +1,5 @@
 const request = require("supertest");
-const app = require("../../src/app");
+const app = require("../../src/config/app.config");
 
 const signupAndLoginUser = async () => {
   const userData = {
