+++ server/test/unfinished/artistController.test.js
@@ -1,4 +1,4 @@
-const app = require("../../src/app");
+const app = require("../../src/config/app.config");
 const User = require("../../src/models/userModel");
 const RefreshToken = require("../../src/models/refreshTokenModel");
 const request = require("supertest");
