+++ server/test/unfinished/userController.test.js
@@ -1,5 +1,5 @@
 const mongoose = require("mongoose");
-const app = require("../../src/app");
+const app = require("../../src/config/app.config");
 const request = require("supertest");
 const User = require("../../src/models/userModel");
 const RefreshToken = require("../../src/models/refreshTokenModel");
