+++ server/test/unit/middlewares/authorize.test.js
@@ -1,7 +1,7 @@
 const authorize = require("../../../src/middlewares/authorize");
 const middlewareMock = require("../../helpers/middlewareMock");
 const User = require("../../../src/models/userModel");
-const app = require("../../../src/app");
+const app = require("../../../src/config/app.config");
 const httpMocks = require("node-mocks-http");
 
 jest.mock("../../../src/models/userModel");
